CLN: ASV frame_ctor benchmarks
